refactor: split conversations.ts into handler modules
