+++ convex/_generated/api.d.ts
@@ -51,13 +51,22 @@ import type * as http from "../http.js";
 import type * as imageModels from "../imageModels.js";
 import type * as internal_ from "../internal.js";
 import type * as lib_capability_resolver from "../lib/capability_resolver.js";
+import type * as lib_conversation_action_handlers from "../lib/conversation/action_handlers.js";
 import type * as lib_conversation_background_operations from "../lib/conversation/background_operations.js";
 import type * as lib_conversation_context_building from "../lib/conversation/context_building.js";
+import type * as lib_conversation_helpers from "../lib/conversation/helpers.js";
 import type * as lib_conversation_message_handling from "../lib/conversation/message_handling.js";
+import type * as lib_conversation_mutation_handlers from "../lib/conversation/mutation_handlers.js";
+import type * as lib_conversation_query_handlers from "../lib/conversation/query_handlers.js";
 import type * as lib_conversation_streaming from "../lib/conversation/streaming.js";
 import type * as lib_conversation_summarization from "../lib/conversation/summarization.js";
 import type * as lib_conversation_types from "../lib/conversation/types.js";
 import type * as lib_conversation_utils from "../lib/conversation_utils.js";
+import type * as lib_message_action_handlers from "../lib/message/action_handlers.js";
+import type * as lib_message_helpers from "../lib/message/helpers.js";
+import type * as lib_message_internal_handlers from "../lib/message/internal_handlers.js";
+import type * as lib_message_mutation_handlers from "../lib/message/mutation_handlers.js";
+import type * as lib_message_query_handlers from "../lib/message/query_handlers.js";
 import type * as lib_model_fetchers from "../lib/model_fetchers.js";
 import type * as lib_model_resolution from "../lib/model_resolution.js";
 import type * as lib_models_dev_mapping from "../lib/models_dev_mapping.js";
@@ -135,13 +144,22 @@ declare const fullApi: ApiFromModules<{
   imageModels: typeof imageModels;
   internal: typeof internal_;
   "lib/capability_resolver": typeof lib_capability_resolver;
+  "lib/conversation/action_handlers": typeof lib_conversation_action_handlers;
   "lib/conversation/background_operations": typeof lib_conversation_background_operations;
   "lib/conversation/context_building": typeof lib_conversation_context_building;
+  "lib/conversation/helpers": typeof lib_conversation_helpers;
   "lib/conversation/message_handling": typeof lib_conversation_message_handling;
+  "lib/conversation/mutation_handlers": typeof lib_conversation_mutation_handlers;
+  "lib/conversation/query_handlers": typeof lib_conversation_query_handlers;
   "lib/conversation/streaming": typeof lib_conversation_streaming;
   "lib/conversation/summarization": typeof lib_conversation_summarization;
   "lib/conversation/types": typeof lib_conversation_types;
   "lib/conversation_utils": typeof lib_conversation_utils;
+  "lib/message/action_handlers": typeof lib_message_action_handlers;
+  "lib/message/helpers": typeof lib_message_helpers;
+  "lib/message/internal_handlers": typeof lib_message_internal_handlers;
+  "lib/message/mutation_handlers": typeof lib_message_mutation_handlers;
+  "lib/message/query_handlers": typeof lib_message_query_handlers;
   "lib/model_fetchers": typeof lib_model_fetchers;
   "lib/model_resolution": typeof lib_model_resolution;
   "lib/models_dev_mapping": typeof lib_models_dev_mapping;
